% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourScores47.R
\name{AImove}
\alias{AImove}
\title{Move of AI}
\usage{
AImove(field, AIstrength, AIplayernumber)
}
\arguments{
\item{field}{matrix: the playing field}

\item{AIstrength}{integer: strength of the AI - number of moves the AI will simulate?}

\item{AIplayernumber}{integer: 0 or 1: should the AI be player 1 or player 2?}
}
\value{
the selected row
}
\description{
Help-Function for an AI
}
