% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourScores47.R
\name{clicking}
\alias{clicking}
\title{a function}
\usage{
clicking(buttons, x, y)
}
\arguments{
\item{buttons}{the mouse buttons input.}

\item{x}{the x-value of the mouse button.}

\item{y}{the y-value of the mouse button.}
}
\value{
a rounded value for the x-coordinate
}
\description{
help-function which return the x-axis-value of the mouse when releasing the mouse button.
}
