% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourScores47.R
\name{clickingXY}
\alias{clickingXY}
\title{check input}
\usage{
clickingXY(buttons, x, y)
}
\arguments{
\item{buttons}{the mouse buttons input.}

\item{x}{the x-value of the mouse button.}

\item{y}{the y-value of the mouse button.}
}
\value{
a Vector of the x and y coordinates of the mouse click
}
\description{
a function to check the mouse click input by the user
}
