% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourScores46.R
\name{fbuttons}
\alias{fbuttons}
\title{Field buttons}
\usage{
fbuttons(field, justsub, message, MACuser, rows, columns, AI, AIstrength,
  PlayerNames, PlayerColors)
}
\arguments{
\item{field}{matrix: the field.}

\item{justsub}{boolean: should only be a subtitle plotted (below the winning field)?}

\item{message}{character: a message to be plotted.}

\item{MACuser}{boolean: on some non-mac computers this can be set to \code{FALSE} to have mouse-functionality in the graphics device.}

\item{rows}{integer: how many rows shall the playing field have?}

\item{columns}{integer: how many columns shall the playing field have?}

\item{AI}{boolean: play against AI?}

\item{AIstrength}{integer: strength of the AI - number of moves the AI will simulate?}

\item{PlayerNames}{array of characters: the players' names.}

\item{PlayerColors}{vector of characters: the players' colors.}
}
\description{
A function to show buttons, letting the player(s) decide what to do: show the winning field, play again or exit.
}
