% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourScores46.R
\name{FieldCorrect}
\alias{FieldCorrect}
\title{Is the field correct?}
\usage{
FieldCorrect(column, field)
}
\arguments{
\item{column}{integer: the column chosen by the current player}

\item{field}{matrix: the playing field.}
}
\value{
a boolean (\code{TRUE} if the given column would be a valid move for the field given).
}
\description{
help-function that checks whether the field is correct
}
