% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourScores46.R
\name{FieldGeneration}
\alias{FieldGeneration}
\title{field generation}
\usage{
FieldGeneration(rows, columns)
}
\arguments{
\item{rows}{integer: how many rows shall the playing field have?}

\item{columns}{integer: how many columns shall the playing field have?}
}
\value{
an empty matrix with rows and columns
}
\description{
help-function which generates the playing-field
}
