% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourScores47.R
\name{FieldPlot}
\alias{FieldPlot}
\title{plot the field}
\usage{
FieldPlot(field, message, PlayerColors)
}
\arguments{
\item{field}{matrix: the playing field}

\item{message}{character: a message to be plotted.}

\item{PlayerColors}{vector of characters: the players' colors.}
}
\description{
a major-function which plots the current field, and if given a hint, which player has won
}
