% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourScores47.R
\name{FieldWinCheck}
\alias{FieldWinCheck}
\title{check for a winner}
\usage{
FieldWinCheck(field, player)
}
\arguments{
\item{field}{matrix: the playing field.}

\item{player}{integer: the current player.}
}
\value{
a boolean whether the player has won the match or not
}
\description{
help-function that checks whether (at least) one of the four possibilities of winning is given
}
