\name{FourScores-package}
\alias{FourScores-package}
\docType{package}
\title{
  FourScores - Who gets first four in a row wins
}
\description{
  A game for two player: Who gets first four in a row wins. Published by Milton Bradley, designed by Howard Wexler and Ned Strongin.
}
\details{
\tabular{ll}{
  Package: \tab FourScores\cr
  Type: \tab Package\cr
  Version: \tab 1.0\cr
  Date: \tab 2011-10-25\cr
  License: \tab GPL (>= 2)\cr
  LazyLoad: \tab yes\cr
}

}
\author{
  Matthias Speidel

  Maintainer: Matthias Speidel <matthias.speidel@googlemail.com>

}
\references{
  Published by Milton Bradley, designed by Howard Wexler and Ned Strongin.
}

