% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourScores46.R
\name{FourScores}
\alias{FourScores}
\title{Main Function}
\usage{
FourScores(rows = 6, columns = 7, AI = FALSE, AIstrength = 14,
  MACuser = TRUE, PlayerNames = c("Player1", "Player2"),
  getnewnames = TRUE, PlayerColors = c("green", "blue"),
  getnewcolors = TRUE)
}
\arguments{
\item{rows}{integer: how many rows shall the playing field have?}

\item{columns}{integer: how many columns shall the playing field have?}

\item{AI}{boolean: play against AI?}

\item{AIstrength}{integer: strength of the AI - number of moves the AI will simulate?}

\item{MACuser}{boolean: on some non-mac computers this can be set to \code{FALSE} to have mouse-functionality in the graphics device.}

\item{PlayerNames}{array of characters: the players' names.}

\item{getnewnames}{boolean: should new names be asked for?}

\item{PlayerColors}{vector of characters: the players' colors.}

\item{getnewcolors}{boolean: should new colors be asked for?}
}
\description{
Function to play FourScores
}
\examples{
\dontrun{
FourScores(AI = T, AIstrength = 10, MACuser = T, getnewnames = F, getnewcolors = F)
}
}
