% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourScores47.R
\name{getColors}
\alias{getColors}
\title{A function}
\usage{
getColors(PlayerNames, PlayerColors, MACuser)
}
\arguments{
\item{PlayerNames}{array of characters: the players' names.}

\item{PlayerColors}{vector of characters: the players' colors.}

\item{MACuser}{boolean: on some non-mac computers this can be set to \code{FALSE} to have mouse-functionality in the graphics device.}
}
\value{
a vector with the updated player colors
}
\description{
A function to get some colors
}
