% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourScores46.R
\name{getPlayerNames}
\alias{getPlayerNames}
\title{Get player names}
\usage{
getPlayerNames(PlayerNames, MACuser)
}
\arguments{
\item{PlayerNames}{array of characters: the players' names.}

\item{MACuser}{boolean: on some non-mac computers this can be set to \code{FALSE} to have mouse-functionality in the graphics device.}
}
\value{
a vector with the player names
}
\description{
help-function which gets and returns the players' names
}
