% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourScores47.R
\name{NewField}
\alias{NewField}
\title{Generate a new field}
\usage{
NewField(field, column, player)
}
\arguments{
\item{field}{matrix: the playing field.}

\item{column}{integer: the column chosen by the current player.}

\item{player}{integer: the current player.}
}
\value{
The updated field matrix.
}
\description{
help-function which "throws" the stone into the field and returns the new field
}
