% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourScores47.R
\name{painter}
\alias{painter}
\title{logo painter}
\usage{
painter(numberMatrix, colorArray)
}
\arguments{
\item{numberMatrix}{a matrix with different integers showing which color to pick from the \code{colorArray}.}

\item{colorArray}{a character array with different names of colors to be used by the painter.}
}
\description{
a general help function to plot
}
