% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourScores46.R
\name{resample}
\alias{resample}
\title{resample}
\usage{
resample(x, ...)
}
\arguments{
\item{x}{a vector}

\item{...}{other parameters}
}
\value{
a vector
}
\description{
resampling function
}
\references{
Help function from \code{?sample} to overcome the "sample(ret, size = 1)" problem for length(ret) == 1
}
