% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourScores47.R
\name{typing}
\alias{typing}
\title{Return a key}
\usage{
typing(key)
}
\arguments{
\item{key}{a keyboard input.}
}
\value{
the key pressed.
}
\description{
help-function which returns, the key on the keyboard which is being typed
}
