% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourScores46.R
\name{AImove}
\alias{AImove}
\title{Move of AI}
\usage{
AImove(field, columns, rows, AIstrength)
}
\arguments{
\item{field}{matrix: the playing field}

\item{columns}{integer: how many columns shall the playing field have?}

\item{rows}{integer: how many rows shall the playing field have?}

\item{AIstrength}{integer: strength of the AI - number of moves the AI will simulate?}
}
\value{
the selected row
}
\description{
Help-Function for an AI
}
